% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity.R
\docType{methods}
\name{clusplotk,ORdensity-method}
\alias{clusplotk,ORdensity-method}
\title{clusplotk}
\usage{
\S4method{clusplotk}{ORdensity}(object, k = object@bestK)
}
\description{
clusplotk
}
\examples{

}
