% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity.R
\docType{methods}
\name{compute.ORdensity,ORdensity-method}
\alias{compute.ORdensity,ORdensity-method}
\title{compute.ORdensity}
\usage{
\S4method{compute.ORdensity}{ORdensity}(object, B = 100, scale = FALSE,
  alpha = 0.05, fold = floor(B/10), weights = c(1/4, 1/2, 1/4),
  K = 10, verbose = FALSE, parallel = FALSE, replicable = TRUE,
  seed = 0)
}
\description{
compute.ORdensity
}
\examples{

}
