% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDE.R
\docType{class}
\name{findDE-class}
\alias{findDE-class}
\alias{findDE}
\title{An S4 class to represent a bank account.}
\description{
An S4 class to represent a bank account.
}
\section{Slots}{

\describe{
\item{\code{positive}}{A length-one numeric vector}

\item{\code{negative}}{A length-one numeric vector}
}}

