% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDE (copy).R, R/findDE.R
\docType{class}
\name{findDE}
\alias{findDE}
\title{Class for representing findDE}
\description{
Class for representing findDE

Class for representing findDE
}
\author{
Itziar Irigoien, Concepcion Arenas, Jose Maria Martinez-Otzeta

Itziar Irigoien, Concepcion Arenas, Jose Maria Martinez-Otzeta
}
