% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity-package.R
\docType{package}
\name{ORdensity-package}
\alias{ORdensity-package}
\alias{ORdensity}
\title{ORdensity}
\description{
ORdensity is a package for the automated discovery of differentially expressed genes.
It makes use of the ORdensity method and the associated FP and dFP values to detect the
most likely true positives.
}
\seealso{
\code{\link{summary.ORdensity}}, \code{\link{preclusteredData}}, 
\code{\link{plot.ORdensity}}, \code{\link{silhouetteAnalysis}},
\code{\link{clusplotk}}, \code{\link{compute.ORdensity}}, 
\code{\link{findbestK}}, \code{\link{findDEgenes}}, \code{\link{ORdensity}}
}
\author{
Jose Maria Martinez Otzeta \email{josemaria.martinezo@ehu.eus}
}
