% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity.R
\docType{class}
\name{ORdensity}
\alias{ORdensity}
\title{Class for representing ORdensity}
\description{
Class for representing ORdensity

An S4 class to represent an ORdensity object.
}
\section{Slots}{

\describe{
\item{\code{positive}}{}

\item{\code{negative}}{}
}}

\author{
Itziar Irigoien, Concepcion Arenas, Jose Maria Martinez-Otzeta
}
