% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity.R
\name{plot.ORdensity}
\alias{plot.ORdensity}
\title{plot.ORdensity}
\usage{
plot.ORdensity(object, k = object@bestK, ...)
}
\arguments{
\item{object}{An object of \code{\link{ORdensity}} class}

\item{K}{Integer value to indicate the maximum number of partitions to check 
along the successive clusterings. By default \code{K=10}.}
}
\value{
Displays a plot of the average silhouette width along with the number of clusters in the partition. 
If assigned to an object, it returns a vector of length \code{K} with the average
silhouette values.
}
\description{
This function plots the \code{\link{silhouette}} values of successive clusterings.
}
\details{
Once the potential DE genes are identified, the real DE genes and the not real DE genes or
false positives must be distinguished and it is carried out clustering  all the potential DE genes  
(see \code{\link{findDEgenes}}). The clustering algorithm is   \code{\link{pam}} and by default
the number of clusters in the partition is obtained by \code{\link{silhouette}}. 
Nevertheless, \code{silhouetteAnalysis} allows to check whether there is a clear best number for the partition
or the  average silhouettes values are
rather similar. 
#' @references Rousseeuw, P. J. (1987). Silhouettes: a graphical aid to the interpretation and validation of cluster analysis. Journal of computational and applied mathematics, 20, 53-65.
}
\examples{
# Read data from 2 experimental conditions
x <- simexpr[, 3:32]
y <- simexpr[, 33:62]
EXC.1 <- as.matrix(x)
EXC.2 <- as.matrix(y)
myORdensity <- new("ORdensity", Exp_cond_1 = EXC.1, Exp_cond_2 = EXC.2)
silhouetteAnalysis(myORdensity)
}
