% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity.R
\docType{methods}
\name{plotFPvsOR,ORdensity-method}
\alias{plotFPvsOR,ORdensity-method}
\title{plotFPvsOR}
\usage{
\S4method{plotFPvsOR}{ORdensity}(object, k = object@bestK)
}
\description{
plotFPvsOR
}
\examples{

}
