% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity.R
\docType{methods}
\name{preclusteredData}
\alias{preclusteredData}
\title{preclusteredData}
\usage{
preclusteredData(object, ...)
}
\arguments{
\item{object}{Object of class ORdensity}
}
\value{
\code{\link{data.frame}} with all potential DE genes
}
\description{
This function returns offers the description of all the identified 
potential DE genes in terms of variables OR, FP, and dFP in one only table.
}
\examples{
randomA <- matrix(rnorm(30*1000), nrow = 1000)
randomB <- matrix(rnorm(30*1000), nrow = 1000)
myORdensity <- new("ORdensity", Exp_cond_1 = randomA, Exp_cond_2 = randomB, parallel = TRUE)
preclusteredData(myORdensity)
}
