% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity.R
\docType{methods}
\name{silhouetteAnalysis,ORdensity-method}
\alias{silhouetteAnalysis,ORdensity-method}
\title{silhouetteAnalysis}
\usage{
\S4method{silhouetteAnalysis}{ORdensity}(object)
}
\description{
silhouetteAnalysis
}
\examples{

}
