% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity.R
\docType{class}
\name{ORdensity}
\alias{ORdensity}
\title{Class for representing ORdensity}
\description{
Class for representing ORdensity
}
\author{
Itziar Irigoien, Concepcion Arenas, Jose Maria Martinez-Otzeta
}
