% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity-package.R, R/ORdensity.R
\docType{package}
\name{ORdensity}
\alias{ORdensity}
\alias{ORdensity-package}
\title{ORdensity}
\description{
ORdensity is a package for the automated discovery of differentially expressed genes.
It makes use of the ORdensity method and the associated FP and dFP values to detect the
most likely true positives.

Object of class ORdensity includes all potential diffentially expressed genes
given miroarray data measured in two experimental conditions.
}
\section{Slots}{

\describe{
\item{\code{positive}}{Matrix including microarray data measured under experimental condition 1.}

\item{\code{negative}}{matrix including microarray data measured under experimental condition 2.}

\item{\code{labels}}{Vector of characters identifying the genes, by default
rownames(positive) is inherited. If NULL,
the genes are named ‘Gene1’, ..., ‘Genen' according to the order given in \code{positive}.}

\item{\code{B}}{Numeric value indicating the number of bootstrap iteration. By default, \code{B}=100.}

\item{\code{scale}}{Logical value to indicate whether the scaling of the difference of quatiles should be done.}

\item{\code{alpha}}{Numeric value to control the bootstrap threshold. By default 0.05.}

\item{\code{fold}}{Numeric value, by default \code{fold}=10. It controls the number of partitions.}

\item{\code{probs}}{Vector of numerics. It sets the quantiles to be considered. By default
\code{probs = c(0.25, 0.5, 0.75)}.}

\item{\code{weights}}{Vector of numerics. It controls the weights given to the quantiles set in \code{probs}. 
By default  \code{weights = c(1/4, 1/2, 1/4)}.}

\item{\code{K}}{Numeric value to set the number of nearest neighbours. By default \code{K=10}.}

\item{\code{out}}{List containing the potential DE genes and their characteristics.
...}
}}

\examples{
# To create an instance of a class ORdensity given data from 2 experimental conditions
x <- simexpr[, 3:32]
y <- simexpr[, 33:62]
EXC.1 <- as.matrix(x)
EXC.2 <- as.matrix(y)
myORdensity <- new("ORdensity", Exp_cond_1 = EXC.1, Exp_cond_2 = EXC.2)
}
\references{
Irigoien and Arenas (2018)
Identification of differentially expressed genes by means of outlier detection.
\emph{BMC Bioinformatics}, 19:317
}
\seealso{
\code{\link{summary.ORdensity}}, \code{\link{preclusteredData}}, 
\code{\link{plot.ORdensity}}, \code{\link{silhouetteAnalysis}},
\code{\link{clusplotk}}, \code{\link{compute.ORdensity}}, 
\code{\link{findbestK}}, \code{\link{findDEgenes}}, \code{\link{ORdensity}}
}
\author{
Jose Maria Martinez Otzeta \email{josemaria.martinezo@ehu.eus}

Jose Maria Martinez-Otzeta, Itziar Irigoien, Concepcion Arenas
}
