% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity.R
\docType{methods}
\name{preclusteredData}
\alias{preclusteredData}
\title{preclusteredData}
\usage{
preclusteredData(object, ...)
}
\arguments{
\item{object}{Object of type ORdensity}
}
\value{
The data before doing any clustering
}
\description{
This function returns the data after applying the ORdensity procedure, but before doing any clustering
}
\examples{
randomA <- matrix(rnorm(30*100), nrow = 100)
randomB <- matrix(rnorm(30*100), nrow = 100)
myORdensity <- new("ORdensity", Exp_cond_1 = randomA, Exp_cond_2 = randomB, parallel = TRUE)
preclusteredData(myORdensity)
}
