% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity.R
\docType{methods}
\name{preclusteredData,ORdensity-method}
\alias{preclusteredData,ORdensity-method}
\title{preclusteredData}
\usage{
\S4method{preclusteredData}{ORdensity}(object)
}
\description{
preclusteredData
}
\examples{

}
