% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORdensity.R
\docType{methods}
\name{preclusteredData}
\alias{preclusteredData}
\title{preclusteredData}
\usage{
preclusteredData(object, ...)
}
\arguments{
\item{object}{Object of class \code{\link{ORdensity}}}
}
\value{
\code{\link{data.frame}} with all potential DE genes
}
\description{
This function returns the description of all the identified 
potential DE genes in terms of variables OR, FP, and dFP in one only table so that 
the interesed user can perform its own clustering analysis.
}
\examples{
# Read data from 2 experimental conditions
x <- simexpr[, 3:32]
y <- simexpr[, 33:62]
EXC.1 <- as.matrix(x)
EXC.2 <- as.matrix(y)
myORdensity <- new("ORdensity", Exp_cond_1 = EXC.1, Exp_cond_2 = EXC.2)
# data.frame with all potential DE genes:
preclusteredData(myORdensity)
}
\seealso{
\code{\link{summary.ORdensity}}, 
\code{\link{plot.ORdensity}}, \code{\link{silhouetteAnalysis}},
\code{\link{findDEgenes}}, \code{\link{ORdensity}}
}
